+++ backend/src/main.py
@@ -5,7 +5,7 @@ import sentry_sdk
 from .models import NewsArticle
 from .config import App
 from .database import SessionLocal
-from .news.utils import get_new
+from .news.services import get_new
 app = FastAPI()
 sentry_sdk.init(
     dsn=App.DSN,
