+++ backend/src/news/services.py
